clean-core: string ordering compares unsigned bytes
cc::string_view::compare did int(_data[i]) - int(other._data[i]) on a char that
is signed on our platforms, so every byte >= 0x80 sorted ahead of all ASCII.
interned_string::compare_bytes inherited it through as_string_view().compare(),
while documenting itself as ordering "by the canonical bytes" and being "the
same in every process". The existing tests only compared "alpha" against "beta",
which cannot tell the two orders apart.

Milestone 1 already hit this on vdoc's object keys and worked around it locally
with impl::compare_key_bytes. Filing it in the caller rather than fixing it here
is why it came back, and the second occurrence is worse: milestone 2 sorts
entity/component/property ids, which are arbitrary application strings, and that
sort feeds the op hash. A document with one non-ASCII entity id would have baked
signed-char ordering into the .vdoc wire format, where a third-party reader
sorting by actual byte value computes a different op_id and neither side can fix
it afterwards.

So compare now widens through u8, matching what the header already promised and
what std::string_view does. compare_key_bytes stays as a thin name over it,
because the value format is defined in terms of that order and a call site should
say so — but it is no longer a second implementation of it.

Full suite passes on all four check presets, so nothing depended on the old
order.

+++ libs/base/clean-core/src/clean-core/string/string_view.hh
@@ -347,13 +347,18 @@ public:
 public:
     /// Lexicographically compares this string_view with another.
     /// Returns: <0 if *this < other, 0 if equal, >0 if *this > other.
+    ///
+    /// Compares by UNSIGNED byte value, so the order is the byte order the content actually has.
+    /// `char` is signed on our platforms, so comparing it directly would sort any byte >= 0x80 before every ASCII one —
+    /// which would put UTF-8 text in an order no other implementation agrees with, this one included once the bytes are
+    /// written to a file or hashed.
     [[nodiscard]] constexpr int compare(string_view other) const
     {
         auto const min_size = _size < other._size ? _size : other._size;
         for (isize i = 0; i < min_size; ++i)
         {
             if (_data[i] != other._data[i])
-                return int(_data[i]) - int(other._data[i]);
+                return int(u8(_data[i])) - int(u8(other._data[i]));
         }
         return _size < other._size ? -1 : (_size > other._size ? 1 : 0);
     }
